Background Music, Bug Fix, and Verticality
I added background music to the game as I felt that it was far too quiet and not exciting enough. I also fixed a glitch that occurred in which the 'endmusic' would only play once and not upon subsequent game overs. I also adjusted the controls and added verticality to the game, as I felt this made it more fun and not as unfair and difficult.