2 Fixes.
Below the surface - what's going on with the player's attack,
this was fixed as the code within the collision file did not perform the required task.

Below the surface - landing next to (but not in) an acid pool damages the player
this was fixed by reducing the size of the toxic pool (this was a cosmetic change) that greatly improves the visual of the pool being in a hole in the ground, the object was also changed to allow for it to remain the same bottom y axis location.